fix token api

+++ Sever/app/Http/Controllers/Ticket/TicketController.php
@@ -23,18 +23,18 @@ class TicketController
     public function getAllTickets()
     {
         try {
-            // $user = Auth::user();
-            // $userId = $user->id;
-            // $pageName = "View Tickets";
+            $user = Auth::user();
+            $userId = $user->id;
+            $pageName = "View Tickets";
 
-            // $permission = $this->permissionBiz->getPermission($pageName, $userId);
+            $permission = $this->permissionBiz->getPermission($pageName, $userId);
             
-            // if ($permission) {
-            //     return response()->json($tickets);
-            // } else {
-            //     return response()->json(['message' => 'Bạn không có quyền xem vé.'], 403);
-            // }
-            $tickets = $this->ticketBusiness->getAllTickets();
+            if ($permission) {
+                $tickets = $this->ticketBusiness->getAllTickets();
+                return response()->json($tickets);
+            } else {
+                    return response()->json(['message' => 'Bạn không có quyền xem vé.'], 403);
+                }
         } catch (\Exception $e) {
             return response()->json(['message' => 'Đã xảy ra lỗi', 'error' => $e->getMessage()], 500);
         }
@@ -44,17 +44,17 @@ class TicketController
     public function createTicket(Request $request)
     {
         try {
-            // $user = Auth::user();
-            // $userId = $user->id;
-            // $pageName = "Add Ticket";
+            $user = Auth::user();
+            $userId = $user->id;
+            $pageName = "Add Ticket";
 
-            // $permission = $this->permissionBiz->getPermission($pageName, $userId);
+            $permission = $this->permissionBiz->getPermission($pageName, $userId);
 
-            // if ($permission) {
-            // } else {
-            //     return response()->json(['message' => 'Bạn không có quyền thêm vé.'], 403);
-            // }
-            $ticketId = $this->ticketBusiness->createTicket($request->all());
+            if ($permission) {
+                $ticketId = $this->ticketBusiness->createTicket($request->all());
+            } else {
+                return response()->json(['message' => 'Bạn không có quyền thêm vé.'], 403);
+            }
             return response()->json(['message' => 'Thêm vé thành công', 'ticket_id' => $ticketId]);
         } catch (\Exception $e) {
             return response()->json(['message' => 'Đã xảy ra lỗi', 'error' => $e->getMessage()], 500);
@@ -65,18 +65,18 @@ class TicketController
     public function updateTicket(Request $request, int $ticketId)
     {
         try {
-            // $user = Auth::user();
-            // $userId = $user->id;
-            // $pageName = "Update Ticket";
+            $user = Auth::user();
+            $userId = $user->id;
+            $pageName = "Update Ticket";
 
-            // $permission = $this->permissionBiz->getPermission($pageName, $userId);
+            $permission = $this->permissionBiz->getPermission($pageName, $userId);
 
-            // if ($permission) {
-            //     return response()->json(['message' => 'Cập nhật vé thành công']);
-            // } else {
-            //     return response()->json(['message' => 'Bạn không có quyền cập nhật vé.'], 403);
-            // }
-            $this->ticketBusiness->updateTicket($ticketId, $request->all());
+            if ($permission) {
+                $this->ticketBusiness->updateTicket($ticketId, $request->all());
+                return response()->json(['message' => 'Cập nhật vé thành công']);
+            } else {
+                return response()->json(['message' => 'Bạn không có quyền cập nhật vé.'], 403);
+            }
         } catch (\Exception $e) {
             return response()->json(['message' => 'Đã xảy ra lỗi', 'error' => $e->getMessage()], 500);
         }
@@ -86,18 +86,18 @@ class TicketController
     public function deleteTicket(int $ticketId)
     {
         try {
-            // $user = Auth::user();
-            // $userId = $user->id;
-            // $pageName = "Delete Ticket";
+            $user = Auth::user();
+            $userId = $user->id;
+            $pageName = "Delete Ticket";
 
-            // $permission = $this->permissionBiz->getPermission($pageName, $userId);
+            $permission = $this->permissionBiz->getPermission($pageName, $userId);
 
-            // if ($permission) {
-            //     return response()->json(['message' => 'Xóa vé thành công']);
-            // } else {
-            //     return response()->json(['message' => 'Bạn không có quyền xóa vé.'], 403);
-            // }
-            $this->ticketBusiness->deleteTicket($ticketId);
+            if ($permission) {
+                $this->ticketBusiness->deleteTicket($ticketId);
+                return response()->json(['message' => 'Xóa vé thành công']);
+            } else {
+                return response()->json(['message' => 'Bạn không có quyền xóa vé.'], 403);
+            }
         } catch (\Exception $e) {
             return response()->json(['message' => 'Đã xảy ra lỗi', 'error' => $e->getMessage()], 500);
         }
